add location to story

+++ lib/features/add_story/domain/entity/add_story_entity.dart
@@ -1,4 +1,3 @@
-import 'dart:ffi';
 import 'dart:io';
 
 import 'package:equatable/equatable.dart';
@@ -6,8 +5,8 @@ import 'package:equatable/equatable.dart';
 class AddStoryEntity extends Equatable {
   final String description;
   final File photo;
-  final Float? lat;
-  final Float? lon;
+  final double? lat;
+  final double? lon;
 
   const AddStoryEntity(this.description, this.photo, this.lat, this.lon);
